+++ Math.cpp
@@ -9,18 +9,18 @@ using namespace std;
 
 Vector vzero(0,0,0);
 
-float squared_distance_between_points(const Vector* p1, const Vector* p2)
+double squared_distance_between_points(const Vector* p1, const Vector* p2)
 {
 	return (p2->x-p1->x)*(p2->x-p1->x) + (p2->y-p1->y)*(p2->y-p1->y) + (p2->z-p1->z)*(p2->z-p1->z);
 }
 
-float cos_at_ab(float a2, float b2, float c2, float ab)
+double cos_at_ab(double a2, double b2, double c2, double ab)
 {
 	return (a2 + b2 - c2) / (2 * ab);
 }
 
 // Works for 0-180
-bool acute_angle(float cos)
+bool acute_angle(double cos)
 {
 	return cos > 0;
 }
@@ -36,11 +36,11 @@ Vector cross_product(const Vector* a, const Vector* b, const Vector* c)
 // http://kb->komires->net/article->php?id=2
 // http://pl->wikipedia->org/wiki/Iloczyn_wektorowy
 // http://www->matematyka->pl/47774->htm
-float triangle_area(const Vector* a, const Vector* b, const Vector* c)
+double triangle_area(const Vector* a, const Vector* b, const Vector* c)
 {
 	Vector axb = cross_product(a, b, c);
 	//cout << a << b << c << axb;
-	float area = 0.5f * sqrt( squared_distance_between_points(&vzero, &axb) );
+	double area = 0.5f * sqrt( squared_distance_between_points(&vzero, &axb) );
 	return area;		
 }
 
@@ -50,7 +50,7 @@ float triangle_area(const Vector* a, const Vector* b, const Vector* c)
 /**
  * Returns true on intersection 
  */
-bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vector *s, float r)
+bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vector *s, double r)
 {	
 	if (p1 == p2)
 	{
@@ -59,16 +59,16 @@ bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vec
 		return squared_distance_between_points(p1, s) <= r*r;
 	}	
 
-	float sq_p1_to_s = squared_distance_between_points(p1, s);
-	float sq_p2_to_s = squared_distance_between_points(p2, s);
-	float sq_p1_to_p2 = squared_distance_between_points(p1, p2);
-	float p1_to_p2 = sqrt( sq_p1_to_p2 );
-	float p1_to_s = sqrt( sq_p1_to_s );
-	float p2_to_s = sqrt( sq_p2_to_s );
+	double sq_p1_to_s = squared_distance_between_points(p1, s);
+	double sq_p2_to_s = squared_distance_between_points(p2, s);
+	double sq_p1_to_p2 = squared_distance_between_points(p1, p2);
+	double p1_to_p2 = sqrt( sq_p1_to_p2 );
+	double p1_to_s = sqrt( sq_p1_to_s );
+	double p2_to_s = sqrt( sq_p2_to_s );
 
 
-	float cos_p1 = cos_at_ab(sq_p1_to_s, sq_p1_to_p2, sq_p2_to_s, p1_to_p2 * p1_to_s );
-	float cos_p2 = cos_at_ab(sq_p2_to_s, sq_p1_to_p2, sq_p1_to_s, p1_to_p2 * p2_to_s );
+	double cos_p1 = cos_at_ab(sq_p1_to_s, sq_p1_to_p2, sq_p2_to_s, p1_to_p2 * p1_to_s );
+	double cos_p2 = cos_at_ab(sq_p2_to_s, sq_p1_to_p2, sq_p1_to_s, p1_to_p2 * p2_to_s );
 
 	 // cout << "dbg" << cos_p1 << " " << acos(cos_p1) * 360 / 2 / M_PI << " " << acute_angle(cos_p1) << endl;
 	 // cout << "dbg" << cos_p2 << " " << acos(cos_p2) * 360 / 2 / M_PI << " " << acute_angle(cos_p2) << endl;
@@ -104,28 +104,28 @@ bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vec
 	return false;
 }
 
-float diffusion_coefficient(float temperature_K, float viscosity_eta, float diameter)
+double diffusion_coefficient(double temperature_K, double viscosity_eta, double diameter)
 {
-	static float k_b = 1.38E-10;
+	static double k_b = 1.38E-10;
 	
 	return k_b * temperature_K / ( 3 * M_PI * diameter * viscosity_eta);
 }
 
-float squared_displacement(int dimensions, float diffusion_coefficient, float tau)
+double squared_displacement(int dimensions, double diffusion_coefficient, double tau)
 {
 	return 2 * dimensions * diffusion_coefficient * tau;
 }
 
 //http://www.johndcook.com/cpp_phi.html
-float phi(float x)
+double phi(double x)
 {
     // constants
-    float a1 =  0.254829592;
-    float a2 = -0.284496736;
-    float a3 =  1.421413741;
-    float a4 = -1.453152027;
-    float a5 =  1.061405429;
-    float p  =  0.3275911;
+    double a1 =  0.254829592;
+    double a2 = -0.284496736;
+    double a3 =  1.421413741;
+    double a4 = -1.453152027;
+    double a5 =  1.061405429;
+    double p  =  0.3275911;
 
     // Save the sign of x
     int sign = 1;
@@ -134,16 +134,16 @@ float phi(float x)
     x = fabs(x)/sqrt(2.0);
 
     // A&S formula 7.1.26
-    float t = 1.0/(1.0 + p*x);
-    float y = 1.0 - (((((a5*t + a4)*t) + a3)*t + a2)*t + a1)*t*exp(-x*x);
+    double t = 1.0/(1.0 + p*x);
+    double y = 1.0 - (((((a5*t + a4)*t) + a3)*t + a2)*t + a1)*t*exp(-x*x);
 
     return 0.5*(1.0 + sign*y);
 }
 
 // approximation
 //http://www.protonfish.com/random.shtml
-float normal(float mean, float dev)
+double normal(double mean, double dev)
 {
-	float x = (((float)rand()/RAND_MAX) * 2 - 1) + (((float)rand()/RAND_MAX) * 2 - 1) + (((float)rand()/RAND_MAX) * 2 - 1);
+	double x = (((double)rand()/RAND_MAX) * 2 - 1) + (((double)rand()/RAND_MAX) * 2 - 1) + (((double)rand()/RAND_MAX) * 2 - 1);
 	return dev * x + mean;
 }
